fix: bf16 CPU compat, clean SFT data gen, push all training scripts
- generate_sft_data.py: rewritten with 15 high-quality pairs, fixed $ne quoting
- train_sft.py: bf16=GPU_AVAILABLE (was fp16, _BaseConfig defaults bf16=True)
- train_dpo_v6.py: bf16=GPU_AVAILABLE (fixes CPU ValueError)
- train_grpo.py: unchanged (CPU fallback works)
- train-lora.yml: unchanged (separate SFT gen/train steps, PR diagnostics)

+++ kin-gguf/train_dpo_v6.py
@@ -250,7 +250,7 @@ def main():
             save_total_limit=2,
             max_length=MAX_LENGTH,
             report_to="none",
-            fp16=GPU_AVAILABLE,
+            bf16=GPU_AVAILABLE,
             gradient_checkpointing=GPU_AVAILABLE,
             optim="adamw_torch",
             lr_scheduler_type="cosine",
